pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.3",
    "google-genai>=1.62.0",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
